bugfix in select, with new test
There was an ordering issue with select, if you selected
columns in an order outside the original, then the types
and rows were mislabeled.

For example, if you had columns A,B, and C, and did a
select for A,C, it would work. If you selected C,A, then
the column types and data rows would be wrong.

Changed the loop in select columns to not depend on the
ordering being the same, which I think will be slightly
slower.